% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rgig}
\alias{rgig}
\title{A Generator for Generalized Gaussian Distribution}
\usage{
rgig(N, lambda, psi, chi, eps = 0.5, K = 0L)
}
\arguments{
\item{N}{number of variates.}

\item{lambda}{parameter lambda.}

\item{psi}{parameter psi.}

\item{chi}{parameter chi.}

\item{eps}{desired rejection rate between 0 and 1; active only when `K` == 0.}

\item{K}{desired number of cutoff points.}
}
\value{
a vector of `N` random variates of the specified GIG distribution.
}
\description{
A function that generates random variates from generalized Gaussian distributions. 
User can specify a rejection rate upper bound or a desired number of cutoff points.
}
\note{
See Zhang and Reiter (2022).
}
